feat: add flow utilities and templates
